+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # Change Log
 
+## [0.0.2] - 2018-06-06
+
+**Fixes**
+
+- Changed `fs.readDirSync` to use the `vscode` api
+- Fixed an issue where only the `lib` files were modified instead of the `src` files
+- Added a check to see if the `strings` import is already present in the current file
+
 ## [0.0.1] - 2018-06-06
 - Initial BETA release